+++ Main/testchem.py
@@ -507,35 +507,33 @@ class Compound(Chemistry):
         """the string representation of a compound object"""
         return str(self.formula)  
         
-def populate(anObject):
-    """"""
+def populate(anObject): 
     if type(anObject) == type([]) or type(anObject) == type(np.zeros(1)):
         try:
             locDict = {}
             for i in range(len(anObject)):
                 try:locDict[i] = int(anObject[i])
-                except (ValueError,TypeError,AttributeError): pass
-            
-            locKeys = locDict.keys();locKeys.sort();locKeys.reverse()
+                except (ValueError, AttributeError,TypeError):anObject[i] = populate(anObject[i])
+                    
+            locKeys = locDict.keys()
             if len(locKeys) == 0:raise Exception
-                
+            locKeys.sort();locKeys.reverse()
+            
             for item in locKeys:
-                anObject[item-1] = populate(anObject[item-1])
-                insertObject = anObject[item-1]
-                aNum = int(anObject.pop(item))
-                if aNum > 1:
-                    for i in range(aNum-1):anObject.insert(i+item,insertObject)
+                amount = locDict[item]
+                anObject.pop(item)
+                for i in range(amount-1):
+                    anObject.insert(item,anObject[item-1])
             return anObject
+                
         except:
-            for i in range(len(anObject)):anObject[i] = populate(anObject[i])
+            for i in range(len(anObject)):
+                anObject[i] = populate(anObject[i])
             return anObject
             
-        for i in range(len(anObject)):anObject[i] = populate(anObject[i])
-        return anObject
-        
-    if type(anObject) == type(''):return anObject 
-    
-    else:return anObject
+    if type(anObject) == type(''):
+        try:return toElement(anObject)
+        except:return anObject
 
 def circulate(aRing,aStruc,mark=''):
     #aRing = stringify(aRing[0])